Moved test class to other user storage classes
