+++ tests/base/src/test/java/org/keycloak/tests/admin/userstorage/ServiceAccountUserStorageTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.tests.federation.storage;
+package org.keycloak.tests.admin.userstorage;
 
 import jakarta.ws.rs.BadRequestException;
 import jakarta.ws.rs.core.Response;
